revise eating

+++ srcs/philo/philo_utils.c
@@ -6,7 +6,7 @@
 /*   By: nsakanou <nsakanou@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 16:33:30 by nsakanou          #+#    #+#             */
-/*   Updated: 2024/02/08 16:19:37 by nsakanou         ###   ########.fr       */
+/*   Updated: 2024/02/21 20:27:03 by nsakanou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,12 +32,13 @@ void	print_message(char *str, t_philo *philo)
 	}
 }
 
-int	ft_usleep(int ms, t_philo *philo)
+int	ft_usleep(unsigned int ms, t_philo *philo)
 {
 	int	start;
 
 	start = get_current_time();
-	while ((get_current_time() - start) < ms && !check_philo_die(philo))
+	while (((unsigned int)(get_current_time() - start))
+			< ms && !check_philo_die(philo))
 		usleep(50);
 	return (0);
 }
